@@ -1,46 +1,43 @@
 {
   "name": "api-rest-laudo-nodejs",
-  "version": "1.0.0",
+  "version": "0.1.1",
+  "type": "module",
   "main": "index.js",
   "repository": "https://github.com/lucas-moura1/api-rest-laudo-nodejs.git",
   "author": "Lucas Moura <lucas3ms.ls@gmail.com>",
   "license": "MIT",
-  "engines": {
-    "node": "14"
-  },
   "scripts": {
-    "build": "babel src --out-dir build",
-    "dev": "nodemon src/index.js --exec babel-node",
-    "start": "node build/index",
+    "dev": "nodemon src/index.js",
+    "start": "node src/index.js",
     "test": "jest"
   },
   "dependencies": {
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
+    "express": "^4.18.1",
     "is-base64": "^1.1.0",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.12.11",
-    "nanoid": "^3.1.23",
-    "winston": "^3.3.3",
-    "yup": "^0.32.9"
+    "mongoose": "^6.6.1",
+    "nanoid": "3.3.4",
+    "winston": "^3.8.2",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
-    "@babel/core": "^7.14.3",
-    "@babel/node": "^7.14.2",
-    "@babel/preset-env": "^7.14.2",
-    "@types/jest": "^26.0.23",
-    "babel-jest": "^27.0.1",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.27.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.3",
+    "@babel/cli": "^7.18.10",
+    "@babel/core": "^7.19.1",
+    "@babel/node": "^7.19.1",
+    "@babel/preset-env": "^7.19.1",
+    "@types/jest": "^29.0.3",
+    "babel-jest": "^29.0.3",
+    "dotenv": "^16.0.2",
+    "eslint": "^8.23.1",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "jest": "^27.0.1",
-    "nodemon": "^2.0.7",
-    "prettier": "^2.3.0"
+    "eslint-plugin-promise": "^6.0.1",
+    "jest": "^29.0.3",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.7.1"
   }
 }
